backend_micro/requirements.txt
# ====================================
# SERVIDOR
# ====================================
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-dotenv>=1.0.0

# ====================================
# BASE DE DATOS
# ====================================
sqlalchemy>=2.0.23
sqlmodel                  # Estructura Política
pymysql                   # MySQL/MariaDB (EP y Mesa de Regalos)
cryptography              # Cifrado para pymysql

# ====================================
# VALIDACIÓN Y CONFIGURACIÓN
# ====================================
pydantic[email]
pydantic-settings         # Estructura Política y Mesa de Regalos

# ====================================
# AUTENTICACIÓN Y SEGURIDAD
# ====================================
google-auth>=2.25.2
pyjwt>=2.8.0
python-jose[cryptography]>=3.3.0   # Mesa de Regalos
passlib[bcrypt]>=1.7.4             # Mesa de Regalos + Estructura Política
msal                               # Microsoft OAuth (Estructura Política)

# ====================================
# EXCEL CONVERTER — PDF y OCR
# ====================================
pdfplumber>=0.11.4
camelot-py>=0.11.0
pymupdf>=1.24.10
pytesseract>=0.3.10
Pillow>=10.0.0

# ====================================
# PROCESAMIENTO DE DATOS
# ====================================
pandas
openpyxl>=3.1.5
python-dateutil>=2.9.0
fpdf2                     # Estructura Política (certificados)

# ====================================
# HTTP Y SCRAPING
# ====================================
httpx>=0.27.0
requests>=2.32.0
beautifulsoup4>=4.12.0    # Mesa de Regalos (scraper Mercado Libre)

# ====================================
# IA / OPENAI
# ====================================
openai>=1.52.0

# ====================================
# PAGOS
# ====================================
paypalrestsdk             # Estructura Política

# ====================================
# EMAIL
# ====================================
aiosmtplib               # Estructura Política (async SMTP)
email-validator

# ====================================
# TESTS
# ====================================
pytest
